hw refactoring
