+++ 2_day/app.js
@@ -13,7 +13,6 @@
  */
 
 import logger from './lib/logger/logger.js';
-
 import color from './color.js';
 import fruit from './fruit.js';
 import { add } from './handler.js';
@@ -25,4 +24,3 @@ log.warn(fruit);
 log.error('ERROR occur: My log');
 
 add(3, 5);
-
