+++ 2_day/lib/logger/appenders/file.js
@@ -1,13 +1,9 @@
 import fs from 'fs';
 import * as constants from '../constants.js';
+import { formatMessage } from '../helper.js';
 
 function log(date, level, category, message) {
-  let fileName = '';
-
-  fileName = constants.logLevel.ERROR === level ? `${category}_error.txt` : `${category}.txt`
-  fs.writeFileSync(fileName, formatMessage(date, level, category, message), {flag: "a+"});
-}
-function formatMessage(date, level, category, message) {
-  return `Date: ${date}, category: ${category}, level: ${level}, message: ${JSON.stringify(message)}\n`;
+  let fileName = constants.logLevel.ERROR === level ? './app_error.log' : './app.log'
+  fs.writeFileSync(fileName, `${formatMessage(date, level, category, message)}\n`, {flag: "a+"});
 }
 export default { log };
