+++ 2_day/lib/logger/config/envConfig.js
@@ -1,0 +1,17 @@
+import * as constants from '../constants.js';
+import { ifValid } from '../helper.js';
+
+export function setEnvConfig() {
+  const config = {};
+
+  for (let prop in constants.customProps) {
+    const property = constants.customProps[prop];
+    const value = process.env[prop].toUpperCase();
+
+    if(value && ifValid(constants[property], value)) {
+      config[property] = process.env[prop].toUpperCase();
+    }
+  }
+
+  return config;
+}
