+++ 2_day/lib/logger/logger.js
@@ -1,6 +1,6 @@
-import config from './config.js';
+import config from './config/config.js';
 import { scoreLevel, logLevel } from './constants.js';
-import * as appenderStrategy from "./appenderStrategy.js"
+import * as appenderStrategy from "./appenders/appenderStrategy.js"
 
 const logger = (category) => ({
   info: (message) => {
